# Advanced buffer benchmark system