… internal/buffer/buffer_generated_test.go