This PR replaces the ad-hoc buffer benchmark from #1526 with a much more sophisticated system that benchmarks creating, reading, and editing separately across many buffer sizes and cursor counts. It then analyzes the results using [benchstat](https://godoc.org/golang.org/x/perf/cmd/benchstat). Running `make bench` gives you the following:

```
name                           time/op
CreateAndClose10Lines-4        63.4µs ±15%
CreateAndClose100Lines-4        142µs ±27%
CreateAndClose1000Lines-4       890µs ± 1%
CreateAndClose10000Lines-4     8.99ms ± 0%
CreateAndClose100000Lines-4    80.7ms ± 0%
CreateAndClose1000000Lines-4    803ms ± 0%
Read10Lines-4                  1.92µs ± 1%
Read100Lines-4                 12.1µs ± 1%
Read1000Lines-4                 120µs ± 1%
Read10000Lines-4               1.37ms ± 1%
Read100000Lines-4              14.3ms ± 2%
Read1000000Lines-4              127ms ± 1%
Edit10Lines1Cursor-4           72.4µs ± 1%
Edit100Lines1Cursor-4          43.5µs ± 0%
Edit100Lines10Cursors-4        3.00ms ± 1%
Edit1000Lines1Cursor-4         24.0µs ± 1%
Edit1000Lines10Cursors-4       2.47ms ± 0%
Edit1000Lines100Cursors-4       193ms ± 0%
Edit10000Lines1Cursor-4         121µs ± 4%
Edit10000Lines10Cursors-4      2.91ms ± 0%
Edit10000Lines100Cursors-4      215ms ± 0%
Edit10000Lines1000Cursors-4     18.1s ± 0%
Edit100000Lines1Cursor-4       1.92ms ± 0%
Edit100000Lines10Cursors-4     35.0ms ± 0%
Edit100000Lines100Cursors-4     556ms ± 0%
Edit100000Lines1000Cursors-4    21.7s ± 0%
Edit1000000Lines1Cursor-4       354µs ± 0%
Edit1000000Lines10Cursors-4     341ms ± 0%
Edit1000000Lines100Cursors-4    3.86s ± 1%
Edit1000000Lines1000Cursors-4   59.1s ± 0%
```

This demonstrates that creating and reading buffers scales linearly in the buffer size, which is as good as it gets. It also shows that *editing* buffers scales (roughly) linearly in the buffer size for a given number of cursors, and superlinearly (quadratically?) in the number of cursors for a given buffer size. The latter is emphatically *not* as good as it gets, and an obvious first target for optimization. Note that buffer texts and editing operations are randomly generated (with fixed seeds) in order to catch "optimizations" that unknowingly exploit benchmark structure, which explains apparent deviations from the asymptotic runtime behavior.

But there is more! Running `make bench-baseline` executes the benchmarks and saves the result in a baseline file. After making changes to the code, you can then simply run `make bench-compare` to get this:

```
name                           old time/op  new time/op  delta
CreateAndClose10Lines-4        55.1µs ±31%  44.7µs ±20%   ~     (p=0.400 n=3+3)
CreateAndClose100Lines-4        121µs ± 7%   134µs ±15%   ~     (p=0.400 n=3+3)
CreateAndClose1000Lines-4       893µs ± 2%   877µs ± 1%   ~     (p=0.400 n=3+3)
CreateAndClose10000Lines-4     8.91ms ± 1%  9.06ms ± 1%   ~     (p=0.200 n=3+3)
CreateAndClose100000Lines-4    80.2ms ± 0%  80.5ms ± 0%   ~     (p=0.400 n=3+3)
CreateAndClose1000000Lines-4    800ms ± 0%   801ms ± 0%   ~     (p=0.400 n=3+3)
Read10Lines-4                  1.92µs ± 2%  1.96µs ± 1%   ~     (p=0.200 n=3+3)
Read100Lines-4                 12.3µs ± 1%  12.3µs ± 1%   ~     (p=1.000 n=3+3)
Read1000Lines-4                 122µs ± 1%   121µs ± 1%   ~     (p=0.700 n=3+3)
Read10000Lines-4               1.39ms ± 3%  1.36ms ± 1%   ~     (p=1.000 n=3+3)
Read100000Lines-4              14.3ms ± 1%  14.1ms ± 0%   ~     (p=0.100 n=3+3)
Read1000000Lines-4              128ms ± 1%   126ms ± 1%   ~     (p=0.200 n=3+3)
Edit10Lines1Cursor-4           72.1µs ± 1%  71.6µs ± 0%   ~     (p=0.200 n=3+3)
Edit100Lines1Cursor-4          43.7µs ± 1%  43.6µs ± 0%   ~     (p=1.000 n=3+3)
Edit100Lines10Cursors-4        3.04ms ± 2%  2.99ms ± 1%   ~     (p=0.100 n=3+3)
Edit1000Lines1Cursor-4         24.3µs ± 1%  24.1µs ± 1%   ~     (p=0.400 n=3+3)
Edit1000Lines10Cursors-4       2.48ms ± 0%  2.47ms ± 0%   ~     (p=0.200 n=3+3)
Edit1000Lines100Cursors-4       192ms ± 0%   191ms ± 0%   ~     (p=0.100 n=3+3)
Edit10000Lines1Cursor-4         121µs ± 4%   118µs ± 0%   ~     (p=0.200 n=3+3)
Edit10000Lines10Cursors-4      2.91ms ± 0%  2.92ms ± 1%   ~     (p=0.700 n=3+3)
Edit10000Lines100Cursors-4      221ms ± 5%   216ms ± 0%   ~     (p=0.700 n=3+3)
Edit10000Lines1000Cursors-4     18.4s ± 4%   18.1s ± 0%   ~     (p=1.000 n=3+3)
Edit100000Lines1Cursor-4       1.93ms ± 0%  1.91ms ± 0%   ~     (p=0.100 n=3+3)
Edit100000Lines10Cursors-4     35.6ms ± 4%  35.1ms ± 0%   ~     (p=0.700 n=3+3)
Edit100000Lines100Cursors-4     564ms ± 2%   555ms ± 0%   ~     (p=0.700 n=3+3)
Edit100000Lines1000Cursors-4    21.7s ± 0%   21.7s ± 0%   ~     (p=1.000 n=3+3)
Edit1000000Lines1Cursor-4       353µs ± 0%   352µs ± 0%   ~     (p=0.400 n=3+3)
Edit1000000Lines10Cursors-4     342ms ± 0%   341ms ± 1%   ~     (p=0.400 n=3+3)
Edit1000000Lines100Cursors-4    3.86s ± 1%   3.86s ± 0%   ~     (p=0.700 n=3+3)
Edit1000000Lines1000Cursors-4   58.9s ± 0%   59.0s ± 0%   ~     (p=1.000 n=3+3)
```

Note how although there are nominal differences between "old" and "new", benchstat correctly determines that they are statistically insignificant by estimating distribution functions based on the outcomes of multiple benchmark runs. :microscope:

With the infrastructure from this PR and #1526 in place, the door is now open for radical experimentation with buffer internals, while having a complete picture of runtime performance, and the certainty that everything is still working as it should.