@p-e-w 2020-03-11
@zyedidia Are you planning to merge this? I have several potential performance improvements in mind that depend on this system for accurate assessment.