@zyedidia 2020-04-02
Thanks for this, it's very cool! I've noticed it takes quite a long time to run the benchmarks, do you think it might be good to decrease the number of test cases?