@ret394 2020-04-09
from https://github.com/zyedidia/micro/pull/1549/files#diff-3292f312a655eb1976bb3bc356bfbeb5 is seems some benchmarks are missing. please make the new system equivalent.