@p-e-w 2020-05-10
@ret394 That is intentional. The new system isn't designed to be equivalent to the old one, but to completely replace it. The existing benchmark was a stop-gap solution created by cobbling together *tests* to form a single monolithic function that can give *some* indication on whether things have improved or slowed down. The new system benchmarks individual operations across a variety of file sizes. This is a much better approach that enables a fine-grained evaluation of buffer performance.

@zyedidia On my machine, the complete benchmark suite takes about 10 minutes to run. There are multiple ways to reduce this time, but each of them comes with certain drawbacks:

First, the main reason it takes so long is that everything is run not once but three times. This seems wasteful, but is absolutely critical for determining whether a measured performance improvement is real or not. benchstat uses the sample variance across multiple runs to evaluate a statistic that can distinguish between significant and insignificant time differences. Benchmark comparisons without statistical analysis are commonly performed for lots of software, but are *essentially worthless.* Changing the make target to run the benchmarks just once brings the time down to only 3 minutes, but without further context the results become much more difficult to interpret. I therefore recommend leaving it at 3 runs, and anyone who is aware of the limitations can still run `go test -bench` manually to get only a single execution.

Second, the `Edit1000000Lines1000Cursors` benchmark takes almost a minute by itself. Removing this test brings down the runtime to 7 minutes for all three runs, or 2 minutes if the triple-run is also removed. But I would consider this a poor solution. This is the one benchmark that clearly demonstrates super-linear operations time scaling. The fact that it takes so long to run should be considered a bug, and removing the benchmark is like removing a test because it fails, instead of fixing the underlying bug.

Which leads to the third approach for making the benchmarks run faster: Making Micro's buffer faster. I have several ideas for doing just that and intend to explore them, although I am quite short on time at the moment so this will have to wait a while.