@ret394 2020-05-23
@p-e-w I didn't belittle synthetic benchmarks or propose to make things worse in general. no harm comes of making the system into a hybrid (abuse of terminology!) to cover all the  known bases. Especially since you will experiment based on its results and don't seem to favor any regressions.

The old ones did certainly seem to cover things in addition to the current system. And if enabled would need to be adjusted a bit.

- Thoughts about an attempt to enable the old system? What would it require?

-----

forgot this.

thanks. have a nice day. keep up the great work. 👏

[ret394](https://github.com/ret394)